switched whole project coords and types to <int> type

+++ headers/entities/geometry/Cell.hpp
@@ -16,7 +16,7 @@ Description:
 
 struct Cell
 {
-    TwoNum_set<float> coords; //this rappresents only position on the plane
+    TwoNum_set<int> coords; //this rappresents only position on the plane
     Pack* pack = nullptr;
 };
