+++ headers/entities/geometry/Grid.hpp
@@ -15,7 +15,7 @@ private:
     Plane *XY, *XZ, *YZ;
 
 public:
-    Grid(ThreeNum_set<float> maxDimensions); 
+    Grid(ThreeNum_set<int> maxDimensions); 
     //Necessario passare solamente il parametro pacco, in quanto la posizione dipende dal valore del centro del pacco
     void setOccupiedCells(Pack* inputPack); 
     void freeOccupiedCells(Pack* inputPack);
